requirements.txt
PyQt5==5.15.9
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
python-dateutil==2.8.2
pyinstaller==5.13.2
